Drop support for Win7 / Server 2008 R2 (#4742)
