+++ stl/src/winapisupp.cpp
@@ -1,33 +1,18 @@
 // Copyright (c) Microsoft Corporation.
 // SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
 
-// clang-format off
-// Prevent clang-format from reordering <AppModel.h> before <Windows.h>
-#include <Windows.h>
-#include <AppModel.h> // for APPMODEL_ERROR_NO_PACKAGE
-#include <internal_shared.h>
-#include <cstdlib>
 #include <yvals.h>
-// clang-format on
 
-#pragma warning(push)
-#pragma warning(disable : 4265) // non-virtual destructor in base class
-#include <wrl/wrappers/corewrappers.h>
-#pragma warning(pop)
+#include <cstdlib>
+#include <internal_shared.h>
+
+#include <Windows.h>
 
 #if !defined(_ONECORE)
 namespace {
 
 // Use this macro for defining the following function pointers
 #define DEFINEFUNCTIONPOINTER(fn_name) decltype(&fn_name) __KERNEL32Function_##fn_name = nullptr
-
-#if !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
-    DEFINEFUNCTIONPOINTER(GetCurrentPackageId);
-#endif // !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
-
-#if _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
-    DEFINEFUNCTIONPOINTER(GetSystemTimePreciseAsFileTime);
-#endif // _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
 
     DEFINEFUNCTIONPOINTER(GetTempPath2W);
 
@@ -43,107 +28,9 @@ namespace {
 
 #if !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
 
-#if !defined(_CRT_APP)
-#if defined(_ONECORE)
-
-namespace {
-    struct HMODULETraits {
-        using Type = HMODULE;
-
-        static bool Close(Type const h) noexcept {
-            return ::FreeLibrary(h) != FALSE;
-        }
-
-        static Type GetInvalidValue() noexcept {
-            return nullptr;
-        }
-    };
-
-    using HMODULEHandle = Microsoft::WRL::Wrappers::HandleT<HMODULETraits>;
-} // unnamed namespace
-
-extern "C" int __crt_IsPackagedAppHelper() noexcept {
-    static wchar_t const* const possible_apisets[] = {
-        L"api-ms-win-appmodel-runtime-l1-1-1.dll", // Windows 8.1+ APISet
-        L"ext-ms-win-kernel32-package-current-l1-1-0.dll", // Legacy APISet
-        L"appmodel.dll" // LNM implementation DLL
-    };
-
-    for (auto& dll : possible_apisets) {
-        HMODULEHandle const apiset(LoadLibraryExW(dll, nullptr, LOAD_LIBRARY_SEARCH_SYSTEM32));
-        if (!apiset.IsValid()) {
-            continue;
-        }
-
-        auto const get_current_package_id =
-            reinterpret_cast<decltype(&GetCurrentPackageId)>(GetProcAddress(apiset.Get(), "GetCurrentPackageId"));
-
-        if (!get_current_package_id) {
-            continue;
-        }
-
-        UINT32 buffer_length = 0;
-        if (get_current_package_id(&buffer_length, nullptr) == ERROR_INSUFFICIENT_BUFFER) {
-            return 1;
-        } else {
-            return 0;
-        }
-    }
-
-    // Either the app is not packaged or we cannot determine if the app is packaged:
-    return 0;
-}
-
-#else // ^^^ defined(_ONECORE) / !defined(_ONECORE) vvv
-
-extern "C" int __crt_IsPackagedAppHelper() noexcept {
-    LONG retValue       = APPMODEL_ERROR_NO_PACKAGE;
-    UINT32 bufferLength = 0;
-
-    IFDYNAMICGETCACHEDFUNCTION(GetCurrentPackageId) {
-        retValue = pfGetCurrentPackageId(&bufferLength, nullptr);
-    }
-
-    if (retValue == ERROR_INSUFFICIENT_BUFFER) {
-        return 1;
-    }
-
-    // If GetCurrentPackageId was not found, or it returned a different error,
-    // then this is NOT a Packaged app
-    return 0;
-}
-
-#endif // ^^^ !defined(_ONECORE) ^^^
-#endif // ^^^ !defined(_CRT_APP) ^^^
-
-// __crtIsPackagedApp() - Check if the current app is a Packaged app
-//
-// Purpose:
-//        Check if the current application was started through a package.
-//        This determines if the app is a Packaged app or not.
-//
-//        This function uses a new Windows 8 API, GetCurrentPackageId, to detect
-//        if the application is deployed via a package.
-//
-// Entry:
-//        None
-//
-// Exit:
-//        TRUE if Packaged app, FALSE if not.
 // TRANSITION, ABI: preserved for binary compatibility
 extern "C" _CRTIMP2 BOOL __cdecl __crtIsPackagedApp() noexcept {
-#ifdef _CRT_APP
-    return TRUE;
-#else // ^^^ defined(_CRT_APP) / !defined(_CRT_APP) vvv
-    static int isPackaged = -1; // Initialize to undefined state
-
-    // If we've already made this check, just return the prev result
-    if (isPackaged < 0) {
-        isPackaged = __crt_IsPackagedAppHelper();
-    }
-
-    return (isPackaged > 0) ? TRUE : FALSE;
-#endif // ^^^ !defined(_CRT_APP) ^^^
+    return FALSE;
 }
 
 #endif // !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
@@ -267,15 +154,9 @@ extern "C" _CRTIMP2 BOOL __cdecl __crtSetFileInformationByHandle(_In_ HANDLE con
 
 #if _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
 
+// TRANSITION, ABI: preserved for binary compatibility
 extern "C" _CRTIMP2 void __cdecl __crtGetSystemTimePreciseAsFileTime(_Out_ LPFILETIME lpSystemTimeAsFileTime) noexcept {
-    // use GetSystemTimePreciseAsFileTime if it is available (only on Windows 8+)...
-    IFDYNAMICGETCACHEDFUNCTION(GetSystemTimePreciseAsFileTime) {
-        pfGetSystemTimePreciseAsFileTime(lpSystemTimeAsFileTime);
-        return;
-    }
-
-    // ...otherwise use GetSystemTimeAsFileTime.
-    GetSystemTimeAsFileTime(lpSystemTimeAsFileTime);
+    GetSystemTimePreciseAsFileTime(lpSystemTimeAsFileTime);
 }
 
 #endif // _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
@@ -304,14 +185,6 @@ extern "C" _Success_(return > 0 && return < BufferLength) DWORD __stdcall __crtG
 static int __cdecl initialize_pointers() noexcept {
     HINSTANCE hKernel32 = GetModuleHandleW(L"kernel32.dll");
     _Analysis_assume_(hKernel32);
-
-#if !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
-    STOREFUNCTIONPOINTER(hKernel32, GetCurrentPackageId);
-#endif // !defined(_CRT_WINDOWS) && !defined(UNDOCKED_WINDOWS_UCRT)
-
-#if _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
-    STOREFUNCTIONPOINTER(hKernel32, GetSystemTimePreciseAsFileTime);
-#endif // _STL_WIN32_WINNT < _WIN32_WINNT_WIN8
 
     // Note that GetTempPath2W is defined as of Windows 10 Build 20348 (a server release) or Windows 11,
     // but there is no "_WIN32_WINNT_WIN11" constant, so we will always dynamically load it
