--- tests/std/tests/P1135R6_atomic_wait_vista/test.cpp
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-
-#include "test_atomic_wait.hpp"
-
-int main() {
-#if defined(_M_IX86) || defined(_M_X64) && !defined(_M_ARM64EC)
-    assert(__std_atomic_set_api_level(__std_atomic_api_level::__has_srwlock) == __std_atomic_api_level::__has_srwlock);
-    test_atomic_wait();
-#endif // defined(_M_IX86) || defined(_M_X64) && !defined(_M_ARM64EC)
-}
